pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auditx"
version = "0.2.0"
description = "Extensible CLI to run performance/security/configuration checks"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
mysql = ["pymysql>=1.1"]
zabbix = ["zabbix-utils>=0.1"]

[project.scripts]
auditx = "auditx.cli:app"

[project.entry-points."auditx.checks"]
# Example external plugin registration (in a plugin package)
# mypkg = "mypkg.checks"

[project.entry-points."auditx.providers"]
# Example external provider registration
# myplugin.providers = "myplugin.providers"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_redundant_casts = true
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
